requirements.txt
flask
flask-cors
gunicorn
python-dotenv
pymupdf
tiktoken
tqdm
langchain-core
langchain-google-genai
langchain-mongodb
pymongo
google-generativeai
flasgger
certifi
dnspython
chromadb
langchain-chroma
sentence-transformers
langchain-huggingface
